Update button functionality in kiln interface
The "New" and "Exit" buttons in the kiln interface had their behavior updated. For the "New" button, the "selected_profile" parameter was removed from the "enterNewMode" function call. For the "Exit" button, the "leaveEditMode" function was removed and replaced with a click handler in kiln_control.js that calls a new "cancelProfileEdit" function.